+++ pkg/apis/common/client_opts.go
@@ -1,0 +1,11 @@
+package common
+
+type ClientOpts struct {
+	Environment Environment
+}
+
+type Environment string
+
+const EnvironmentDev = Environment("dev")
+const EnvironmentStage = Environment("stage")
+const EnvironmentProd = Environment("prod")