+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * kubernetes: add support for updating the state of a cluster
 * kubernetes: add support for managing clusters
 * kubernetes: add support for kkp project and cluster id and vlans
+* kubernetes: add support for managing nodepools
 
 <!--
 Please add your changelog entry under this comment in the correct category (Security, Fixed, Added, Changed, Deprecated, Removed - in this order).